Move files into separate sub-packages
Build issues still exist at this point, mostly just invalid package imports.
